% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{brute_force_lower_ci}
\alias{brute_force_lower_ci}
\title{Brute-force computation of confidence interval.}
\usage{
brute_force_lower_ci(
  x_vec,
  bruth_force_ci_helper,
  ci_upper_const = 100,
  tol = 1e-06
)
}
\arguments{
\item{x_vec}{A vector of observations}

\item{bruth_force_ci_helper}{R function used to compute the lower bound of confidence interval.}

\item{ci_upper_const}{Constant to compute upper bound of the width of confidence interval of the form \code{ci_upper_const / sqrt(n)}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
Lower bound of confidence interval given the sample size.
}
\description{
Given a vector of observations, compute the lower bound of confidence interval at the last index.
}
