% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stcp_ci.R
\name{build_bf_ci_bounded}
\alias{build_bf_ci_bounded}
\title{Build a brute-force generator of lower bound of always-valid confidence intervals for bounded random variables.}
\usage{
build_bf_ci_bounded(
  alpha,
  n_upper,
  n_lower,
  bound_lower,
  k_max = 200,
  tol = 1e-06
)
}
\arguments{
\item{alpha}{ARL parameter in (0,1)}

\item{n_upper}{Upper bound of the target sample interval}

\item{n_lower}{Lower bound of the target sample interval. Default is \code{1}.}

\item{bound_lower}{Lower bound of observations.}

\item{k_max}{Positive integer to determine the maximum number of baselines. Default is \code{200}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
\code{stcp_bf_ci} object
}
\description{
Build a brute-force generator of lower bound of always-valid confidence intervals
for bounded random variables given target range of numbers of observations.
}
