% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stcp_ci.R
\name{build_ci_exp}
\alias{build_ci_exp}
\title{Build a generator of lower bound of always-valid confidence intervals for simple exponential e-values.}
\usage{
build_ci_exp(
  alpha,
  n_upper,
  n_lower,
  psi_fn_list_generator = generate_sub_G_fn,
  v_min = 1,
  k_max = 200,
  tol = 1e-06
)
}
\arguments{
\item{alpha}{ARL parameter in (0,1)}

\item{n_upper}{Upper bound of the target sample interval}

\item{n_lower}{Lower bound of the target sample interval. Default is \code{1}.}

\item{psi_fn_list_generator}{Function operator generating \code{psi_fn_list}}

\item{v_min}{A lower bound of v function in the baseline process. Default is \code{1}.}

\item{k_max}{Positive integer to determine the maximum number of baselines. Default is \code{200}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
\code{STCP_CI} Object
}
\description{
Build a generator of lower bound of always-valid confidence intervals
for simple exponential e-values given target range of numbers of observations.
}
