% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_edcp_model.R
\name{build_edcp_bounded}
\alias{build_edcp_bounded}
\title{Build EDCP model for bounded random variables.}
\usage{
build_edcp_bounded(
  alpha,
  m_pre,
  delta_lower,
  delta_upper = bound_upper - m_pre,
  is_test = FALSE,
  bound_lower = 0,
  bound_upper = 1,
  k_max = 1000,
  tol = 1e-06,
  var_lower = 0,
  var_upper = 0.25
)
}
\arguments{
\item{alpha}{ARL parameter in (0,1)}

\item{m_pre}{Upper bound of mean of pre-change observations. Must be between \code{bound_lower} and \code{bound_upper}.}

\item{delta_lower}{Lower bound of target Delta. It must have be positive and smaller than or equal to \code{delta_upper}.}

\item{delta_upper}{Upper bound of target Delta. It must have be positive and larger than or equal to \code{delta_lower}.}

\item{is_test}{A Boolean to indicate whether this model is for a sequential test or not.}

\item{bound_lower}{Lower bound of observations.}

\item{bound_upper}{Upper bound of observations.}

\item{k_max}{Positive integer to determine the maximum number of baselines. Default is \code{1000}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}

\item{var_lower}{Lower bounds of variance of scaled post-change observations. Default is \code{0}.}

\item{var_upper}{Upper bounds of variance of scaled post-change observations. Default is \code{0.25}.}
}
\value{
\code{EDCP} Object of 1. Model parameters, 2. Memory for log e-detectors / values, 3. Auxiliaries
}
\description{
Build EDCP model for bounded random variables to detect up-crossing mean-shift.
}
\examples{
build_edcp_bounded(0.01, 0.5, 0.1, 0.4)
}
