% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stcp_model.R
\name{build_stcp}
\alias{build_stcp}
\title{Build STCP model.}
\usage{
build_stcp(alpha, m_pre, is_test, omega, lambda, log_base_fn_generator, ...)
}
\arguments{
\item{alpha}{ARL parameter in (0,1)}

\item{m_pre}{Upper bound of mean of pre-change observations}

\item{is_test}{A Boolean to indicate whether this model is for a sequential test or not.}

\item{omega}{Mixing weights for baseline processes.}

\item{lambda}{Parameters of baseline processes.}

\item{log_base_fn_generator}{Function factory generating log baseline functions.}

\item{...}{arguments to be passed to \code{log_base_fn_generator}.}
}
\value{
\code{STCP} object of 1. Model parameters, 2. Memory for log e-detectors / values.
}
\description{
Build STCP model to detect up-crossing mean-shift.
}
