% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stcp_model.R
\name{build_stcp_exp}
\alias{build_stcp_exp}
\title{Build STCP model for simple exponential e-detectors.}
\usage{
build_stcp_exp(
  alpha,
  m_pre,
  delta_lower,
  delta_upper,
  is_test = FALSE,
  psi_fn_list = generate_sub_G_fn(sig = 1),
  s_fn = function(x) {     x - m_pre },
  v_fn = function(x) {     1 },
  v_min = 1,
  k_max = 200,
  tol = 1e-06
)
}
\arguments{
\item{alpha}{ARL parameter in (0,1)}

\item{m_pre}{Upper bound of mean of pre-change observations}

\item{delta_lower}{Lower bound of target Delta. It must have be positive and smaller than or equal to \code{delta_upper}.}

\item{delta_upper}{Upper bound of target Delta. It must have be positive and larger than or equal to \code{delta_lower}.}

\item{is_test}{A Boolean to indicate whether this model is for a sequential test or not.}

\item{psi_fn_list}{A list of R functions that computes psi and psi_star functions. Can be generated by \code{generate_sub_G_fn()} or counterparts for sub_B and sub_E.}

\item{s_fn}{R function that compute the sum process given an observation.}

\item{v_fn}{R function that compute the variance process given an observation.}

\item{v_min}{A lower bound of v function in the baseline process. Default is \code{1}.}

\item{k_max}{Positive integer to determine the maximum number of baselines. Default is \code{200}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
\code{STCP} object of 1. Model parameters, 2. Memory for log e-detectors / values, 3. Auxiliaries
}
\description{
Build STCP model for simple exponential e-detectors to detect up-crossing mean-shift.
}
