% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stcp_model.R
\name{combine_stcp}
\alias{combine_stcp}
\title{Combine two stcp objects for a given weight.}
\usage{
combine_stcp(obj1, obj2, w = 0.5)
}
\arguments{
\item{obj1}{First \code{stcp} object to combine.}

\item{obj2}{Second \code{stcp} object to combine.}

\item{w}{Weight scalar to determine the importance of the first \code{stcp} object. Default is 0.5.}
}
\value{
Combined \code{stcp} object of 1. Model parameters, 2. Memory for log e-detectors / values.
}
\description{
Combine two stcp objects for a given weight. Only \code{stcp} objects of the same \code{is_test} can be combined.
Once combined, any additional details in input objects will be discarded.
}
