% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stcp.R
\name{combine_stcp_run}
\alias{combine_stcp_run}
\title{Combine two stcp runs for a given weight.}
\usage{
combine_stcp_run(run1, run2, w = 0.5)
}
\arguments{
\item{run1}{First \code{stcp_run} object to combine.}

\item{run2}{Second \code{stcp_run} object to combine.}

\item{w}{Weight scalar to determine the importance of the first \code{stcp_run} object. Default is 0.5.}
}
\value{
Combined \code{stcp_run} object of 1. logarithm of mixture of e-detectors, 2. the first stopped point and 3. Underlying stcp objects.
}
\description{
Combine two stcp runs for a given weight. Only \code{stcp_run} objects of the same \code{is_test} can be combined.
}
