% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_baseline.R
\name{compute_baseline}
\alias{compute_baseline}
\title{Compute baseline processes}
\usage{
compute_baseline(
  alpha,
  delta_lower,
  delta_upper,
  psi_star = function(x) {     x^2/2 },
  psi_star_div = function(x) {     x },
  psi_star_inv = function(y) {     sqrt(2 * y) },
  v_min = 1,
  k_max = 100,
  tol = 1e-06
)
}
\arguments{
\item{alpha}{ARL parameter in (0,1)}

\item{delta_lower}{Lower bound of target Delta. It must have be positive and smaller than or equal to \code{delta_upper}.}

\item{delta_upper}{Upper bound of target Delta. It must have be positive and larger than or equal to \code{delta_lower}.}

\item{psi_star}{R function which computes the convex conjugate of the psi function. Default is \code{function(x)x^2/2}.}

\item{psi_star_div}{R function which computes the derivative of \code{psi_star}. Default is \code{function(x)x}.}

\item{psi_star_inv}{R function which computes the inverse of \code{psi_star}. Default is \code{function(y)sqrt(2*y)}.}

\item{v_min}{A lower bound of v function in the baseline process. Default is \code{1}.}

\item{k_max}{Positive integer to determine the maximum number of baselines. Default is 100.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
A list of 1. Parameters of baseline processes, 2. Mixing weights, 3. Auxiliary values for computation.
}
\description{
Compuate parameters to build baseline processes.
}
\examples{
compute_baseline(0.01, 1, 2)
}
