% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_baseline.R
\name{compute_baseline_bounded}
\alias{compute_baseline_bounded}
\title{Compute baseline processes for e-detectors of bounded random variables.}
\usage{
compute_baseline_bounded(
  alpha,
  m_pre,
  delta_lower,
  delta_upper = bound_upper - m_pre,
  bound_lower = 0,
  bound_upper = 1,
  k_max = 1000,
  tol = 1e-06,
  delta_lower_explicit = NULL,
  delta_upper_explicit = NULL
)
}
\arguments{
\item{alpha}{ARL parameter in (0,1)}

\item{m_pre}{Upper bound of mean of pre-change observations. Must be between \code{bound_lower} and \code{bound_upper}.}

\item{delta_lower}{Lower bound of target Delta. It must have be positive and smaller than or equal to \code{delta_upper}.}

\item{delta_upper}{Upper bound of target Delta. It must have be positive and larger than or equal to \code{delta_lower}.}

\item{bound_lower}{Lower bound of observations.}

\item{bound_upper}{Upper bound of observations.}

\item{k_max}{Positive integer to determine the maximum number of baselines. Default is \code{1000}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}

\item{delta_lower_explicit}{Specified delta_lower replacing the default formula.}

\item{delta_upper_explicit}{Specified delta_upper replacing the default formula.}
}
\value{
A list of 1. Mixing weights, 2. log baseline functions, 3. ARL parameter
}
\description{
Build baseline processes for e-detectors of bounded random variables to detect up-crossing mean-shift.
}
\examples{
compute_baseline_bounded(0.01, 0.5, 0.1, 0.4)
}
