% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci_utils.R
\name{compute_baseline_for_sample_size}
\alias{compute_baseline_for_sample_size}
\title{Compute baseline parameters given target sample size bounds.}
\usage{
compute_baseline_for_sample_size(
  alpha,
  n_upper,
  n_lower,
  psi_fn_list = generate_sub_G_fn(),
  v_min = 1,
  k_max = 200,
  tol = 1e-06
)
}
\arguments{
\item{alpha}{ARL parameter in (0,1)}

\item{n_upper}{Upper bound of the target sample interval}

\item{n_lower}{Lower bound of the target sample interval. Default is \code{1}.}

\item{psi_fn_list}{A list of R functions that computes psi and psi_star functions. Can be generated by \code{generate_sub_G_fn()} or counterparts for sub_B and sub_E.}

\item{v_min}{A lower bound of v function in the baseline process. Default is \code{1}.}

\item{k_max}{Positive integer to determine the maximum number of baselines. Default is \code{200}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
A list of baseline parameters to build \code{ci_helper}
}
\description{
Given target sample size bounds interval for confidence sequences, compute  baseline parameters.
}
