% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_bf_ci}
\alias{compute_bf_ci}
\title{Compute lower bounds of confidence sequence for simple exponential e-values.}
\usage{
compute_bf_ci(
  x_vec,
  stcp_bf_ci_obj,
  width_upper = 100,
  ci_lower_trivial = -Inf,
  max_num_ci = 100,
  tol = 1e-06
)
}
\arguments{
\item{x_vec}{A vector of observations}

\item{stcp_bf_ci_obj}{stcp_bf_ci object, output of \code{build_bf_ci_bounded}}

\item{width_upper}{Constant to compute upper bound of the width of confidence interval of the form \code{width_upper / sqrt(n)}.}

\item{ci_lower_trivial}{A trivial lower bound of confidence interval. Default is \code{-Inf}.}

\item{max_num_ci}{The maximum number of CI to computes. Recomend to set it be less than 100.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
List of 1. running sample mean, 2. sample size vector, 3. lower bounds of confidence sequence.
}
\description{
Given a vector of observations, compute lower bounds of confidence sequence.
}
