% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_brute_force_ci}
\alias{compute_brute_force_ci}
\title{Brute-force computation of confidence interval.}
\usage{
compute_brute_force_ci(
  x_vec,
  bruth_force_ci_helper,
  width_upper = 100,
  ci_lower_trivial = -Inf,
  tol = 1e-06
)
}
\arguments{
\item{x_vec}{A vector of observations}

\item{bruth_force_ci_helper}{R function used to compute the lower bound of confidence interval.}

\item{width_upper}{Constant to compute upper bound of the width of confidence interval of the form \code{width_upper / sqrt(n)}.}

\item{ci_lower_trivial}{A trivial lower bound of confidence interval. Default is \code{-Inf}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
Lower bound of confidence interval given the sample size.
}
\description{
Given a vector of observations, compute the lower bound of confidence interval at the last index.
}
