% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci}
\alias{compute_ci}
\title{Compute lower bounds of confidence sequence for simple exponential e-values.}
\usage{
compute_ci(
  x_vec,
  stcp_ci_obj,
  width_upper = 100,
  ci_lower_trivial = -Inf,
  n_custom_vec = NULL,
  x_bar_pre = 0,
  n_pre = 0,
  tol = 1e-06
)
}
\arguments{
\item{x_vec}{A vector of observations}

\item{stcp_ci_obj}{stcp_ci object, output of \code{build_ci_exp}}

\item{width_upper}{Constant to compute upper bound of the width of confidence interval of the form \code{width_upper / sqrt(n)}.}

\item{ci_lower_trivial}{A trivial lower bound of confidence interval. Default is \code{-Inf}.}

\item{n_custom_vec}{A vector of number of observations for each entry of \code{x_vec}. If NULL, each observation is assumed to be equal to one.}

\item{x_bar_pre}{Sample mean of previous observations. Default is \code{0} (No pre-samples)}

\item{n_pre}{Number of pre-samples. Default is \code{0}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
List of 1. running sample mean, 2. sample size vector, 3. lower bounds of confidence sequence.
}
\description{
Given a vector of observations, compute lower bounds of confidence sequence.
}
