% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci_width}
\alias{compute_ci_width}
\title{Compute confidence sequence for simple exponential e-values.}
\usage{
compute_ci_width(
  x_vec,
  ci_helper,
  ci_upper_const = 100,
  x_bar_pre = 0,
  n_pre = 0,
  tol = 1e-06
)
}
\arguments{
\item{x_vec}{A vector of observations}

\item{ci_helper}{R function from \code{compute_baseline_simple_exp} used to compute the width.}

\item{ci_upper_const}{Constant to compute upper bound of the width of confidence interval of the form \code{ci_upper_const / sqrt(n)}.}

\item{x_bar_pre}{Sample mean of previous observations. Default is \code{0} (No pre-samples)}

\item{n_pre}{Number of pre-samples. Default is \code{0}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
List of 1. running sample mean, 2. sample size vector, 3. widths of confidence sequence.
}
\description{
Given a vector of observations, compute widths of confidence sequence.
}
