% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_single_ci}
\alias{compute_single_ci}
\title{Compute the lower bound of confidence interval for simple exponential e-values.}
\usage{
compute_single_ci(n, x_bar, ci_helper, width_upper = 100, tol = 1e-06)
}
\arguments{
\item{n}{Number of samples}

\item{x_bar}{Sample mean of \code{n} observations}

\item{ci_helper}{R function from \code{compute_baseline_simple_exp} used to compute the width.}

\item{width_upper}{Constant to compute upper bound of the width of confidence interval of the form \code{width_upper / sqrt(n)}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
A lower bound of confidence interval given the sample size.
}
\description{
Given the number of samples n, compute the lower bound of confidence interval for simple exponential e-values.
}
