% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_single_ci_width}
\alias{compute_single_ci_width}
\title{Compute the width of confidence interval for simple exponential e-values.}
\usage{
compute_single_ci_width(n, ci_helper, ci_upper_const = 100, tol = 1e-06)
}
\arguments{
\item{n}{Number of samples}

\item{ci_helper}{R function from \code{compute_baseline_simple_exp} used to compute the width.}

\item{ci_upper_const}{Constant to compute upper bound of the width of confidence interval of the form \code{ci_upper_const / sqrt(n)}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
Width of confidence interval given the sample size.
}
\description{
Given the number of samples n, compute the width of confidence interval for simple exponential e-values.
}
