% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{convert_time_to_delta_bound}
\alias{convert_time_to_delta_bound}
\title{Convert target sample interval to delta bounds}
\usage{
convert_time_to_delta_bound(
  alpha,
  n_upper,
  n_lower,
  psi_fn_list = generate_sub_G_fn(),
  v_min = 1,
  k_max = 200,
  tol = 1e-06
)
}
\arguments{
\item{alpha}{ARL parameter in (0,1)}

\item{n_upper}{Upper bound of the target sample interval}

\item{n_lower}{Lower bound of the target sample interval. Default is \code{1}.}

\item{psi_fn_list}{A list of R functions that computes psi and psi_star functions. Can be generated by \code{generate_sub_G_fn()} or counterparts for sub_B and sub_E.}

\item{v_min}{A lower bound of v function in the baseline process. Default is \code{1}.}

\item{k_max}{Positive integer to determine the maximum number of baselines. Default is \code{200}.}

\item{tol}{Tolerance of root-finding, positive numeric. Default is 1e-6.}
}
\value{
\code{delta_lower}, \code{delta_upper} in compute_baseline for CI computation.
}
\description{
Given a vector of observations, compute widths of confidence sequence.
}
