% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDCP.R
\name{edcp}
\alias{edcp}
\title{E-Detector based Change-Point detection}
\usage{
edcp(
  x_vec,
  baseline_obj,
  prev_log_e_vec = numeric(length(baseline_obj$log_base_fn_list)),
  is_SR_type = TRUE,
  is_test = FALSE
)
}
\arguments{
\item{x_vec}{Observations}

\item{baseline_obj}{Baseline object returned by compute_baseline functions.}

\item{prev_log_e_vec}{A vector of logarithms of mixtures of e-detectors in the previous step. Default is \code{numeric(length(log_base_fn_list))}.}

\item{is_SR_type}{A Boolean to indicator whether the process use SR-type update. If not, the function uses the CUSUM-type update instead of SR-type one.}

\item{is_test}{A Boolean to indicate whether to compute sequential test or not.}
}
\value{
logarithm of mixture of e-detectors and the first stopped point.
}
\description{
Run mixtures of e-detectors to detect the change-point.
}
