% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci_utils.R
\name{generate_ci_helper}
\alias{generate_ci_helper}
\title{Generate ci_helper function for CI computation.}
\usage{
generate_ci_helper(
  alpha,
  omega,
  lambda,
  psi_fn_list_generator,
  is_psi_depend_on_m = FALSE
)
}
\arguments{
\item{alpha}{Confidence level in (0,1)}

\item{omega}{A vector of omega parameters from \code{compute_baseline}, which is a mixing weights.}

\item{lambda}{A vector of lambda parameters from \code{compute_baseline}, which is a set of model parameters.}

\item{psi_fn_list_generator}{Function operator generating \code{psi_fn_list}}

\item{is_psi_depend_on_m}{Indicator whether the psi function depends on the target parameter.}
}
\value{
\code{ci_helper} function to compute confidence intervals
}
\description{
Given baseline parameters, generate ci_helper function for CI computation.
}
