% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_detectors.R
\name{generate_log_base_fn}
\alias{generate_log_base_fn}
\title{Generate log of baseline function}
\usage{
generate_log_base_fn(
  lambda,
  psi_fn = function(x) {
     x^2/2
 },
  s_fn = function(x) {
     x
 },
  v_fn = function(x) {
     1
 }
)
}
\arguments{
\item{lambda}{Lambda parameter of the target baseline function.}

\item{psi_fn}{R function that compute \eqn{\psi(\lambda)}.}

\item{s_fn}{R function that compute the sum process given an observation.}

\item{v_fn}{R function that compute the variance process given an observation.}
}
\value{
A function compute log of baseline process given an observation.
}
\description{
Generate logarithm of the baseline function given lambda parameter, psi_star, sum and variance functions
}
