% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_detectors.R
\name{generate_log_bounded_base_fn}
\alias{generate_log_bounded_base_fn}
\title{Generate log of baseline function for bounded RVs}
\usage{
generate_log_bounded_base_fn(lambda, m = 0.5, bound_lower = 0)
}
\arguments{
\item{lambda}{Lambda parameter of the target baseline function.}

\item{m}{Mean parameter of the target baseline function. It must be strictly larger than \code{bound_lower}.}

\item{bound_lower}{Lower bound of observations. Default is \code{0}.}
}
\value{
A function compute log of baseline process given an observation.
}
\description{
Generate logarithm of the baseline function for bounded RVs in \eqn{[0,1]} given lambda and mean parameters
}
\examples{
generate_log_bounded_base_fn(1)
generate_log_bounded_base_fn(0.5)
}
