% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generator}
\alias{generator}
\title{Generate a vector of simulated observations}
\usage{
generator(max_sample = 1000L, v = 500L, pre_sampler, post_sampler)
}
\arguments{
\item{max_sample}{Total number of simulated observations}

\item{v}{change-point}

\item{pre_sampler}{R function to generate pre-change observations.}

\item{post_sampler}{R function to generate post-change observations.}
}
\value{
A vector of simulated observations
}
\description{
Generate a simulated observations for given pre- and post-observation samplers.
}
