% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{plot.edcp_out}
\alias{plot.edcp_out}
\title{Plot observed e-value / e-detectors}
\usage{
\method{plot}{edcp_out}(
  x,
  y = seq_along(x$log_mix_e_vec),
  draw_detect_line = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{edcp_out object}

\item{y}{A numeric vector of time indices}

\item{draw_detect_line}{If true then draw a vertical line to indicate the detection time.}

\item{add}{Indicator whether to make an overlay on the existing plot or not}

\item{...}{Graphical parameters can be passed except \code{type} and \code{ylab}, which are pre-defined.}
}
\description{
Plot observed e-value / e-detectors
}
