% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{plot.stcp_run}
\alias{plot.stcp_run}
\title{Plot observed e-value / e-detectors}
\usage{
\method{plot}{stcp_run}(
  x,
  n = seq_along(x$log_mix_e_vec),
  draw_detect_line = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{stcp_run object}

\item{n}{A numeric vector of time indices}

\item{draw_detect_line}{If true then draw a vertical line to indicate the detection time.}

\item{add}{Indicator whether to make an overlay on the existing plot or not}

\item{...}{Graphical parameters can be passed except \code{type} and \code{ylab}, which are pre-defined.}
}
\description{
Plot observed e-value / e-detectors
}
