% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.EDCP}
\alias{print.EDCP}
\title{Print EDCP model summary}
\usage{
\method{print}{EDCP}(x, ...)
}
\arguments{
\item{x}{EDCP object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print EDCP model summary
}
