% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.EDCP_BF_CI}
\alias{print.EDCP_BF_CI}
\title{Print EDCP_BF_CI model summary}
\usage{
\method{print}{EDCP_BF_CI}(x, ...)
}
\arguments{
\item{x}{edcp_bf_ci object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print EDCP_BF_CI model summary
}
