% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.EDCP_RUNNED}
\alias{print.EDCP_RUNNED}
\title{Print EDCP run summary}
\usage{
\method{print}{EDCP_RUNNED}(x, ...)
}
\arguments{
\item{x}{edcp_out object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print EDCP run summary
}
