% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.stcp}
\alias{print.stcp}
\title{Print stcp model summary}
\usage{
\method{print}{stcp}(x, ...)
}
\arguments{
\item{x}{stcp object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print stcp model summary
}
