% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.stcp_ci}
\alias{print.stcp_ci}
\title{Print stcp_ci model summary}
\usage{
\method{print}{stcp_ci}(x, ...)
}
\arguments{
\item{x}{stcp_ci object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print stcp_ci model summary
}
