% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.stcp_run}
\alias{print.stcp_run}
\title{Print stcp run summary}
\usage{
\method{print}{stcp_run}(x, ...)
}
\arguments{
\item{x}{stcp_out object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print stcp run summary
}
