% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.STCP_RUNNED}
\alias{print.STCP_RUNNED}
\title{Print STCP run summary}
\usage{
\method{print}{STCP_RUNNED}(x, ...)
}
\arguments{
\item{x}{stcp_out object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print STCP run summary
}
