% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_edcp.R
\name{run_edcp}
\alias{run_edcp}
\title{E-Detector based Change-Point detection and sequential tests.}
\usage{
run_edcp(new_x, edcp_obj, is_SR_type = TRUE)
}
\arguments{
\item{new_x}{A list or vector of observations}

\item{edcp_obj}{EDCP object. It must contains model parameters (omega, log_base_fn_list) and previous log e-detector / values.}

\item{is_SR_type}{A Boolean to indicator whether the process use SR-type update. If not, the function uses the CUSUM-type update instead of SR-type one.}
}
\value{
logarithm of mixture of e-detectors, the first stopped point and updated edcp object.
}
\description{
Run mixtures of e-detectors / e-values to detect the change-point or run a sequential test.
}
