% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stcp.R
\name{run_stcp}
\alias{run_stcp}
\title{E-Detector based Change-Point detection and sequential tests.}
\usage{
run_stcp(new_x, stcp_obj, is_SR_type = TRUE)
}
\arguments{
\item{new_x}{A list or vector of observations}

\item{stcp_obj}{STCP object. It must contains model parameters (omega, log_base_fn_list) and previous log e-detector / values.}

\item{is_SR_type}{A Boolean to indicator whether the process use SR-type update. If not, the function uses the CUSUM-type update instead of SR-type one.}
}
\value{
logarithm of mixture of e-detectors, the first stopped point and updated stcp object.
}
\description{
Run mixtures of e-detectors / e-values to detect the change-point or run a sequential test.
}
