% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_detectors.R
\name{update_log_e_detector}
\alias{update_log_e_detector}
\title{Log of SR- or CUSUM-type e-detectors}
\usage{
update_log_e_detector(
  x_current,
  prev_log_e = 0,
  compute_log_baseline = function(x) {     x - 0.5 },
  is_SR_type = TRUE
)
}
\arguments{
\item{x_current}{The current observation}

\item{prev_log_e}{A numeric value of the mixture of e-detectors in the previous step. Default is \code{0}.}

\item{compute_log_baseline}{R function that compute the log of baseline process based on each observation.}

\item{is_SR_type}{A Boolean indicator whether the process use SR-type update. If not, the function uses the CUSUM-type update instead of SR-type one.}
}
\value{
Updated logarithm of SR- or CUSUM-type e-detector.
}
\description{
Compute logarithm of SR- or CUSUM-type e-detector based on the current observation and previous value.
}
