% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_of_e_detectors.R
\name{update_log_mix_e}
\alias{update_log_mix_e}
\title{Log of mixtures of e-values and SR- or CUSUM-type e-detectors}
\usage{
update_log_mix_e(
  new_x_list,
  weight_vec,
  log_base_fn_list,
  prev_log_e_vec = rep(-Inf, length(log_base_fn_list)),
  is_test = FALSE,
  is_SR_type = TRUE
)
}
\arguments{
\item{new_x_list}{A list of new observations. If each observation is univariate then it can be a vector instead of list of single observations.}

\item{weight_vec}{A vector of mixing weights.}

\item{log_base_fn_list}{R function that compute the log of baseline process based on each observation.}

\item{prev_log_e_vec}{A vector of logarithms of mixtures of e-detectors in the previous step.}

\item{is_test}{A Boolean to indicate whether this model is for a sequential test or not.}

\item{is_SR_type}{A Boolean to indicator whether the process use SR-type update. If not, the function uses the CUSUM-type update instead of SR-type one.}
}
\value{
Updated logarithm of mixture of e-values, SR- or CUSUM-type e-detectors and a vector of each component for the next iteration.
}
\description{
Compute logarithm of mixtures of e-values, SR- or CUSUM-type e-detectors based on the current observation and previous value.
}
