% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_of_e_values.R
\name{update_log_mix_e_values}
\alias{update_log_mix_e_values}
\title{Log of mixtures of e-values for sequential test}
\usage{
update_log_mix_e_values(
  new_x_vec,
  weight_vec,
  log_base_fn_list,
  prev_log_e_vec = numeric(length(log_base_fn_list))
)
}
\arguments{
\item{new_x_vec}{A vector of new observations.}

\item{weight_vec}{A vector of mixing weights.}

\item{log_base_fn_list}{R function that compute the log of baseline process based on each observation.}

\item{prev_log_e_vec}{A vector of logarithms of mixtures of e-values in the previous step.}
}
\value{
Updated logarithm of mixture of e-values and a vector of each component for the next iteration.
}
\description{
Compute logarithm of mixtures of e-values based on the current observation and previous value.
}
