% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{plot.edcp_out}
\alias{plot.edcp_out}
\title{Plot observed e-value / e-detectors}
\usage{
\method{plot}{edcp_out}(edcp_out, n = seq_along(edcp_out$log_mix_e_vec), add = FALSE, ...)
}
\arguments{
\item{edcp_out}{edcp_out object}

\item{n}{A numeric vector of time indices}

\item{add}{Indicator whether to make an overlay on the existing plot or not}
}
\description{
Plot observed e-value / e-detectors
}
