% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.STCP}
\alias{print.STCP}
\title{Print STCP model summary}
\usage{
\method{print}{STCP}(x, ...)
}
\arguments{
\item{x}{STCP object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print STCP model summary
}
