% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.STCP_BF_CI}
\alias{print.STCP_BF_CI}
\title{Print STCP_BF_CI model summary}
\usage{
\method{print}{STCP_BF_CI}(x, ...)
}
\arguments{
\item{x}{stcp_bf_ci object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print STCP_BF_CI model summary
}
