% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_plot.R
\name{print.STCP_CI}
\alias{print.STCP_CI}
\title{Print STCP_CI model summary}
\usage{
\method{print}{STCP_CI}(x, ...)
}
\arguments{
\item{x}{stcp_ci object}

\item{...}{Placeholder for formatting options - not yet implemented.}
}
\description{
Print STCP_CI model summary
}
