% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_quick_simulation}
\alias{run_quick_simulation}
\title{Run a quick simulation for debuging}
\usage{
run_quick_simulation(x_vec, v, baseline_star, baseline_mix)
}
\arguments{
\item{x_vec}{Simulated observations}

\item{v}{change-point}

\item{baseline_star}{baseline object for the delta_star}

\item{baseline_mix}{baseline object for the mixture case}
}
\value{
Stopped point for single and mixtures of SR- and CUSUM-type e-detectors.
}
\description{
Run single and mixtures of e-val, e-detectors for given simulation setting.
}
