% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_quick_simulation}
\alias{run_quick_simulation}
\title{Run a quick simulation for debuging}
\usage{
run_quick_simulation(x_vec, edcp_mix, v = Inf, edcp_star = NULL)
}
\arguments{
\item{x_vec}{Simulated observations}

\item{edcp_mix}{edcp object for the mixture case}

\item{v}{Change-point}

\item{edcp_star}{edcp object for the delta_star (optional)}
}
\value{
Stopped point for single and mixtures of SR- and CUSUM-type e-detectors.
}
\description{
Run single and mixtures of e-val, e-detectors for given simulation setting.
}
